Shortcut removal PR suggestions
